feat: filter namespace watcher
The notify watcher takes a filter which allows the composing of
triggers over a given term within a single watcher. It's then up
to the mapper to compose the change events together.

By changing the namespace watcher to the same as the notify watcher
we can align the two underlying watchers in the same way.