chore: refactor deployControllerCharm method
The deployControllerCharm method is hard to modify as it relies on logic in defer statements to be executed provided bootstrap is successful. This is compounded by the fact that logic is split for CAAS and IAAS deployments.
This change tries simplify the method by removing the use of logic in defer statements and reducing the length of the method.