feat: add ValidateEnterScope and updates
Update EnterScope args to use a relation UUID and unit name. The
relation unit is not created until first use of EnterScope, thus cannot
be used in it's call.

ValidateEnterScope is a new method to check permutations of relation
scope and subordinate status. These determine if the unit can enter
scope.

Added some method heading comments.