feat: use relation service for entering and leaving scope
Replace use of state for relations with the relation service. Expected
errors will remain the same.

NetworksForRelation functionality, dealing with the relationNetworksC,
has been removed for now. It will be addressed in the CMR domain epic.

Unit tests have been implemented and the legacy ones removed.
TestEnterScopeIgnoredForInvalidPrincipals has not been reimplemented as
it tests functionality outside of the scope of the facade.