feat: application state and service for export
Adds a export migration application method to both service and
state. This method will start to service as a way to add more
fields to enable true migrations from dqlte. This will ensure that
we're no longer running from state.

This is a breaking change, migration of complex applications will
break.