chore: drop status export from application domain
And clean up the unused code.

We have just added a new status domain, and moved over the application
and unit status code to it. However, we did not move model migration
code.

Drop the export code from the application domain. This should be handled
by the status domain. That will be the next step