feat: handle charm extra-bindings as an endpoint binding
Application's have two types of endpoints defined in its charm's
metadata: relations and extra-bindings. Add the
application_extra_endpoint to hold mappings of the extra endpoint to a
space.

And applications endpoints will then be a combination of rows from the
application_extra_endpoint and application_endpoint tables.

Update the charm_extra_binding table to include a UUID, for the new link
table.