feat: implement state layer for exposed endpoints
This patch implements the methods of the state layer for exposed
endpoints in the application domain.

Also, the `MergeExposeSettings()` method on the service layer was
updated, adding relevant checks and default values when an empty space
and CIDR slices are provided.

It must be noted that the wildcard endpoint required special cases in
the state layer code, the reason for this is that we have referential
integrity with respect to the endpoints defined in the charm_relation
table, which doesn't contain the wildcard (empty string "") endpoint for
obvious reasons.