feat(migration): check all unit statuses in one go in precheck
Originally, we would pull all of the statuses out of the service status
at a time, and check them all in the migration code. This was very
suboptimal.

Simplify this by pushing this logic into the status domain. Pull out all
statuses at once with a single query, then in the service check the
statuses.