chore: remove fingerprint from resource store storage key
During the developement of the resources domain, a decision was made
that to update the revision or origin of a resource, a new resource UUID
should be generated and the application_resource table should be updated
to point to that new UUID. Prior to this, the resource row was updated
in place.

Due to the old architecuture, when storing resources in the
resource store, the key had to include the hash of the resource, so that
the different versions of resources with the same UUID could be
differentiated.

This is no longer needed. Once a blob is linked to a resource UUID, no
other blobs should be linked to that UUID, if this is attempted, an
error should be returned.

For this reason, this PR removes the fingerprint from the storage key.
It also removes the code in StoreResource that removes the old version
of the resource, since this is now done by the update resource
functions.

This change is motivated by a bug where the fingerprint of some
container image resources can differe depending on the source of the
blob, even when the contents are the same. Storing the hash in key comes
with an implicit assumption that the hash is the same at all points in
the blobs lifetime, which is not necesserily true due to this bug.