test: fix failing pki test in gating
This commit attempts to fix a failing test in the pki suite that was
comparing rsa private keys when they have been created to when they have
been unmarshalled from a pem block.

The reason for the failure is that newer go versions don't use the CRT
value's anymore in RSA and so they are made without any optimisations.
The field is also deprecated.