feat: adds the sshsession worker
This worker is resonsible for watching connections coming in to the controller's
SSH server and then initating a connection back to the controller creating a tunnel
that the controller can use to initiate another (legitimate) SSH connection down.

This worker then pipes the connection to the local SSHD agent. Additionally, it handles
the addition and deletion of ephemeral public keys, received in the watch updates.