feat: align to logger writer
To better emphasise the usefulness of the logsink, it must also
be a logsink for the LogWriter interface. All logs must go through
this type, it ensures that we're writing to just one file. If there
are multiple writers writing to the same file, there is a chance
for corruption. We see that already in the logsink.log file, there
are multiple things writing to the same file and it causes problems.

We will need to have the inverse of this, so that we can read from
the same file at good safe points (full lines).