fix: split model migration status message
The status message in model migration is causing the watcher to
fire if it is updated. This causes a request amplification because
the more migration minions there are, the more messages that are
written. Depending on how many of those notifications in the watcher
get coalesced, depends on how many new notification change occurs.
Not all messages incurr a new watcher change, but it causes at
least one.

The change is to split out the migration status message to its own
collection. The status message document will have the same ID as
the status document ID, so lookup is really simple. Any changes to
the status message document does not impact the changes to the
status document, thus the reduction of watcher triggering changes.

The complication is model summaries and the lack of tests at the
state level. As we're using the same ID, we can ensure that we're
looking at the same status migration document to back fill the
data.