feat(migration): export statuses during migration
We recently added a new domain to handle statuses. We also removed the
export status capability from the original application domain.

Re-implement and re-write status exporting in the new status domain. Use
new, purpose built status service methods to achieve this.

NOTE: We have retained importing the status in the application domain,
since statuses are set on application and unit creation. This may need
some rethinking in the future, but we aren't there yet.