fix: prevent retry of successful phase
If the migration phase has already been tried and it succeeded, then
we don't need to retry it again.

This is caused during a large migration, a minion will be told to
enter the validation phase. Upon being succeeding, and whilst waiting
for the next phase to be dispatched, another validation phase will
be sent. This occurs, when the migration master worker writes a
status message to the migration status doc, which in turn triggers
another update. All minions will then re-run the validation phase.
This causes a compounding effect, where each minion will then keep
attempting to contact the controller. The minion reports it was
either successful or not. If it was not successful, it causes the
minion to crash the worker, thus when it came back the migration
would be over! As the success phase is the point of no return and
there isn't a SUCCESS_DONE phase, the migration master assumes
that all minions completed their task and wrote the agent config.
The validation phase was completed at some point successfully.

The solution here, is to *only* allow the retry of the validation
phase if it was unsucessful. This applies to all phases for the
same migration id. If the migration id has changed, we allow the
re-run of the whole flow.

To ensure that there is no unbounded growth of the processed phases
map, we clean up the processed map if the migration is either not
running or is terminal at the end. The restarting of the worker will
also clean this up.