refactor(status): use app name instead of id in service signature
It turns out the every caller of these status service methods holds an
app name, which is converted to an id via the application service. This
can be simplified dramatically by changing the method signature to use
an app name instead, and convert to an id in the implementation.

In a few cases, this has meant we can drop the application service from
a facade entirely