refactor: remove child loggers use labels
Removes all the child loggers from in the services package. It was
just creating a LOT of noise. The domains aren't separated as clear
boundaries of work, they all interplay together. Thus to actually
see how an application was deployed you would need to join a lot of
namespaces together to understand the logging a whole. The solution
is what we already have to hand, is logger tags.

This wires it up when we create the services and it's set everytime.