feat: refactor upgrade facade to use set tools
This commit refactors the upgrade facade to now define the logic
internally for the SetTools facade call instead of using the common
facade.

While here I have taken the oppurtunity to invert the dependency tree
for creating a new upgrader facade. The code doesn't rely on so much in
the creation now and instead take's it's direct dependencies.