refactor: change timings
If we only receive one validation change request, then we need to keep
retrying until we successfully connect to the target controller, or we fail.
Constantly, retrying at a shorter interval is not a good idea, as it can
cause the target controller to be flooded with requests (DDoS). The default
time to wait for the migration validation to occur is 15 minutes. This time
can be changed by the user. The retry time should be increased exponentially,
along with additional jitter, to avoid flooding the target all at the same
time.

This updates the minion strategy to take into account that we will get
less notifications, so we need to try for longer.