feat(relation): add structured Key type
The relation key is a natural key of a relation. It can be used to find
the relation in the database via the relation service. Currently, an
unstructured string is passed to the database, this is parsed, and used
to find the relation.

This is not ideal as there are no guarnetees on the structure of a
string. Ideally we would know already by this point that it contained
the expected information. If the string was malformed, the best place to
fail would be upon reciving it at the facade.

By encoding the Key as a structured type, we can guarentee that it
contains the expected information, and make it easy to access.

This change adds the new structured key type and replaces current usages
of it with the new structured form.

In the 3.6, if the key was generated in state, then if there are two
relations then, the first is the provider and the second the requirer,
and if only one, then the endpoint has the role peer. The structured Key
enforces that assumption by storing the role informtion explicitly,
rather than just implicitly via the ordering.

Confusingly, there are places in the code where the "relation key" comes
as an argument passed by the user, and therefore may not follow the
rules above. These keys should remain as strings and not be used with
the relation key type, it should only be used when the key is going to
be used in the service.