feat: wire up logsink
This wires up the logsink to write all the controller logs just to
the underlying logger. We don't need to write the controller logs
via the apiserver. If we have a lot of logs being sent to the
controller and the controller is also writing a lot of logs, then
we can DDoS ourselves.

This is just asinine.