refactor(status): remove ptr to StatusInfo in params
For legacy reasons, all our CRUD methods for application/unit status in
the status service and status layer used a pointer to StatusInfo in
their signature.

This turned out to be very inconvenient. Remove the pointers in method
signatures

There is only one logical change that results from this. In the status
service, in a few places we used a nil pointer in place of an unset
status. We can no longer do this.

So I have refactored the state method GetUnitCloudContainerStatus to
return an Unset status if no unit container status is found. Unset is a
new CloudContainerStatusType, so the schema & encode/decode methods
needed to be updated.