feat: return ssh wire format public key
There was a misunderstanding about what state returned in the original PR in that it returned the
public key pre-pem encoded. Instead it returns the private key. This commit parses the key and
marshals the public key into SSH wire format. We've moved away from PEM encoded result to wire
format for ease of use on the client side.