chore: remove dump db
This removes dump db which used mongo as the backend. We no longer
support that. We'll provide another strategy to perform the dump,
probably via the db-repl.

Dumping a model is allowed using model export, but it can not be
simplified. We're not introducing the simplify logic in the model
migration.