feat(config): wire up application config to client application facade
The client application facade has a couple of endpoints which get and
set application/charm config/settings.

Use methods from the application service to cut over these endpoints.

This took a pretty serious refactor of the application domain, as the
existing method to write the application config was not up to purpose.
Instead of overwriting the entire config, we now instead update
individual provided fields.