refactor: remove migration agent version checks
We have a number of pre-checks in model migration around agent version
for unit, application and machine. Because these are currently being
moved over to DQlite some of the calls are failing as they expect to be
coming from Monog in this check.

I have disabled these tests for the time being as there will be future
work to re-enable correct migration for agent binary version.