refactor: remove state based model calls
This change removes the need for the agent/agent facade to acess mongo
state to retrieve the model tag for the current context. Instead we now
pull this information from the facade model context which in turn is
backed by DQlite.

In doing this work we were also able to invert some of the dependencies into the facade constructor removing the need to have an error return from new*.