refactor: move set tools to model agent domain
This commit moves the code we originally had in the unit and machine
service for setting the running version on behalf of an agent binary
into the model agent domain.

The decision was made to group similar operations like these ones into
the their own domain around the same theme over polutting the entity
domain with other concerns.