chore: add wildcard endpoint name to core
This commit adds a file on core/network to keep the wildcard endpoint
name as a constant instead of having it on each domain.
The usages of the constant WildcardEndpoint in domain/port were
therefore replaced by the constant in core/network.

Also, this commit adds tests to the service layer of domain/application
for exposed endpoints.