refactor: agent binary store service layer
This change refactors some of the work done in the agent binary store by
introducing a new set of strict error types for certain error
conditions.

We are also removing the overwrite functionality of the contract so that
if you add a binary that already exists you get an error instead of
overwriting an already existing agent binary that could be wrong or
change behaviour.

This is a concern of the caller to decided what happens in this case. We
should just faciliate the information and look at adding a remove
operation in the future.