chore: merge pull request #19311
https://github.com/juju/juju/pull/19311

If the call to WatchControllerInfo() fails, the client connection is not closed. Given the parent worker retries, if there's an ongoing issue, this will then continue to leak connections.

Also fixes the use of bare channels in the connent and watch method.

The unit tests were rewritten to use go mocks.

unit tests check the client is closed on error

**Jira card:** [JUJU-7754](https://warthogs.atlassian.net/browse/JUJU-7754)

[JUJU-7754]: https://warthogs.atlassian.net/browse/JUJU-7754?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ