refactor: upgrader facade on to model agent
This commit refactors the upgrader facades over to using the model agent
domain for setting reported agent version for machine's and units.

We have left controller node in it's own domain for now as we don't want
to make the model agent have to understand a concept that is outside of
its scope.