feat(relation): implement EnterScope
Implement enter scope for units.

The primary functionality of this is straightforward, however, checking
that the unit is in fact the subordinate unit of the application in the
relation makes it complicated. This requires that the unit_principal
table is populated, which doesn't yet appear to be touched. This may be
an issue later on when it comes to testing, however non subordinates
units are ok.

The change adds several errors, not all of these are useful to the
caller, only the errors are they can use are included in the function
docs.