feat: add timer to removal worker
Passes the clock through to the worker, and uses it to create a timer
based on a maximum interval between check for jobs.

This is useful, because some jobs will simply check for certain
conditions, and do nothing if those are not met. They need to be woken
up to look again.