refactor(status): export status utility functions
StatusInfo and it's associated types & enums are used across a couple of
domains. Previously, these types and their encode/decode functions were
duplicated across these domains.

Chaneg this, by instead exproting these tools in the status domain, and
importing them in others.

Duplicated code, a priori, is NOT necessarily a problem. A little
duplication is better than a little dependency. However, these enums are
not a nice duplication. If one changes and another doesn't, this will
lead to nasty bugs and confusion.

This is particularly obvious for the encode/decode functions, which map
between the status types, and their IDs in the database. Inconsistencies
across copies of this logic would obviously be very bad.