fix(schema): enforce NOT NULL for charm_relation fields role_id and scope_id
There is no reason to allows NULL fields for role_id and scope_id. A relation should always have
a scope and a role to be able to relates to another. This commit set this constraint in database,
replace LEFT JOIN by simple JOIN in related view and update unit test helpers to insert relation in
state.

- `domain/relation/watcher_test.go`: Updated `addCharmRelation` to include `scope_id` and `role_id`
   in the SQL `INSERT`.
- `domain/relation/state/relation_test.go`: Updated `addCharmRelationWithDefaults` to include
  `scope_id` and `role_id` in the SQL `INSERT`.
- `domain/schema/model/sql/0015-charm.sql`: Modified `charm_relation` table to make `scope_id` and
  `role_id` fields mandatory (`NOT NULL`). Updated JOIN clauses to change `LEFT JOIN` to `JOIN` for
  related fields.
- `domain/application/state/application_endpoint_test.go`: Updated SQL `INSERT` in transactional
  test to include mandatory `scope_id` and `role_id` fields.