chore: add proxy for SSH sessions to machines
This change adds logic to connect a user's session with a session to their target machine.
The plumbing is relatively straightforward thanks to the Go x/crypto/ssh package capabilities.
We only need to connect the stdin, stdout, and stderr of both ends and wait for the operation to complete.