fix: return correct error message objectstore
The objectstore was returning the incorrect error message if the
same metadata written multiple times. The solution is to remember
if we inserted the objecstore metadata and if we didn't then return
the hash already exists based on the primary key constraint error.
Otherwise, we can assume that the path already exists with a
different hash.