fix(state): provide resolved charmUUID to insertApplicationEndpoint
Before this fix, the charmID passed as parameter for insertApplicationEndpoint was the
one passed as parameter to AddApplication, which may not be the right one, if a charm
already exists.
After this fix, the correct charmID is given, either the one which just have been inserted or the
one which already exists in the database.