feat: add agent binaries during bootstrap
This commit introduces the new ability for the bootstrap worker to now
upload agent binaries into the controller's agent binary store during
bootstrap.

Previously this was done only in Mongo's similar binary store but now
with this commit we dual write to both places.

The dual writing can be removed once the rest of agent binaries are
using this new store.