refactor: introduces new watchable service
This commit introduces a new watchable service to domain/modelagent.
Specifically with the intention of breaking out the parts of the
original service that dealt with watching.

This was done so that model migration can use the model agent service to
import and export agent version. This can't happen when the watcher is
required on the service.

As part of this commit I have re-generated all of the mocks that mock
out the domain/services.