refactor: rework model manager
Reworks the model worker manager, so that it doesn't request the
services for a model on every restart of the worker. If the worker
is thrashing and restarting, we don't want to keep getting the
domain services. It causes other parts of the system to thrash.
Let the model worker thrash on it's own.

It would help if the following worker patch[1] was landed, as we
could tell the model worker to back off when this happens.

 1. https://github.com/juju/worker/pull/41