feat(relation): remove call to relation state from status facade
The processApplicationRelations function called through to state to get
the endpoints that were already in the relations map.

This change gets the endpoint belonging to the application by iterating
throught the endpoints in place to find the one belonging to the
application and then getting all applications it is related to that have
an endpoint with a counterpart role. If the application is a subordinate
and the endpoint has scope Container, then it is added to the set of
subordinates.

This function could be optimised further, but since we are going to
rewrite status anyway I've made this very light touch.