feat(status): re-wire application status for failed unit hooks
We do something very odd in Juju 3. If a hook fails, the unit agent goes
into an error state. However, when we display this, we show the workload
status to be in error, and the agent status as idle.

In a previous commit, we accounted for this in the unit display
statuses. However, the same was needed for the application status.

This means, if an application status is not explicitly set, the status
is a (quite complex!) function of the workload, agent and container
statuses, as well as the presence of all the units.

This commit implements this function, using the statuses exposed to the
service in a previous commit.

But I have also refactored the state methods to return args consolidated
into a single return value, and improve the names.