chore: remove DeviceConstraints() method from legacy state
This patch removes the `DeviceConstraints()` method from the legacy
state/application.go.
The value was being set in ProvisioningInfo{} for k8s but was
surprisingly never read.

Since this was never used, the DDL for charm was updated, as well as the
charm metadata structs in domain and internal/charm.