feat(relation): implement SetRelationStatus on status service
Implement SetRelationStatus on the status service. This includes
checking that the transition from the old status to the new is allowed
as was done in Relation.SetStatus in 3.6.

The SetStatus method also includes a edge case that was previously
handled in the uniter facade. If going from Suspending to Suspended,
keep the previous status message.