refactor: move watchers to a new file
Anticipating the implemention of WatchLifeSuspendedStatus watchers to
a new file. It will be large and complex.

Updated the signature of WatchLifeSuspendedStatus to take a unit name
rather than an application ID. The service method will determine if the
unit is a subordinate or not and watch different applications based on
the answer. All of the logic from the SubordinateRelationWatcher in the
uniter facade will be pushed into the domain.