refactor(relation): table relation_unit refers relation_endpoint instead of relation
This change introduce a direct reference from a relation unit to the application it belongs,
instead of requiring a JOIN through unit<application_uuid>application_endpoint.

This simplifies the request GetRelationApplicationEndpointName, but make the insert a bit more
complicated (insertion is done through relationUnitUUID) as long as few other requests.

However, new JOIN are unidirectional, which make the overall reasoning about schema simpler.

- `domain/relation/state/relation.go`: Updated queries to use `relation_endpoint_uuid` instead of
  `relation_uuid`. Adjusted JOINs and conditions accordingly.
- `domain/relation/state/relation_test.go`: Updated test cases to align with the new data model,
  replacing `relation_uuid` with `relation_endpoint_uuid`.
- `domain/relation/state/types.go`: Added `RelationEndpointUUID` field in `relationUnit`
  and reflected changes in the data structure.
- `domain/schema/model/sql/0024-relation.sql`: Modified `relation_unit` table schema to replace
  `relation_uuid` with `relation_endpoint_uuid`. Adjusted constraints, indexes and views