refactor: rename machine name to uuid state method
This commit renames the state method used in the model agent service for
fetching a machine's uuid from a corresponding name. This was done so
that the wording of the func name is inline with the unit func used in
the same package.

These methods were coppied from different domains at the start hence the
naming difference from the copy.