feat: two new machine agent version views
This commit introduces two new views to the machine ddl. The first one
is for simply retrieving what has been previously been set on the
machine agent version table but with the added benefit of containing the
arch string which is what a caller really wants to know.

The second view hides the fact that the target agent version for a
machine comes from the model agent version table and also that an entry
needs to exist in the machine agent version table. These are side
affects of poor design in Juju where we don't record the intended
architecture that a machine should be provisioned with.

By introducing the view we get smaller sql in the state files and it
makes it easier later on to break this bad coupling of what should a
machine be running and what is it running.