feat: implement missing state func
This commit implements missing state func
GetMachineRunningAgentBinaryVersion that was previously defined in the
service layer.

This new func directly reports back what has been set by the machine
agent as it reports upwards to the controller. We need to access this
value again whem migrating the model the machine lives in.

Includes tests of the contract.