feat: two new views for unit agent version
This commit introduces two new views on the unit ddl. One for seeing
what unit agent version has already been set for a unit but with the
convenience of also getting the architecture name with no join. This
allows retrieval functions in state be much smaller.

The second view does a similar job to what was done in the machine ddl
by providing a means for getting the target agent version for a unit but
not having to make the caller concerned with how this target value is
currently obtained.

This commit also adds the views for machine and units to the schema
tests.