refactor: make unit target agent binary version use view
This commit changes the unit target agent binary version funcs to now
use the view that compose all of the information together instead of
having to do it inside the func.

This has also allowed the code to now firm up it's contract offer a unit
not found error like it should have.