feat: introduces export operations agent tools
Adds export operations to set agent tools on model migration for
machines and units. This uses previously done work about getting back
the reported agent version from these entities in the model agent
domain.

We avoid setting agent binary parth and sha on export because we have
never maintained a correlation between these two in the model. To report
this information would be to incorrectly guess.