refactor: remove migration tests
This commit removes to migration pre checks that existed for when the
service fails for model agent fails in unrecoverable ways. Specifically
this test was just asserting that an error is passed upwards with no
annotations.

From a testing perspective it is brittle and doesn't assert any buisness
logic. It is likely to break often with error annotations for no reason.