refactor: add back in unit and machine agent check
This commit adds back in checks that had been commented out around unit
and machine agent checks. We disabled these checks while state was being
moved over from Mongo to DQlite and can now re-enable them now that we
are turning on model migration again.

These new checks use the model agent service to query the agent version
for models and units.

The way we have done these checks is a little awkward as we have to go
to the database every time for the information for each machine. In an
ideal world we could wrap up this summary inside the database within a
single transaction.