refactor: migration pre check for agent version
When moving entity tool version over to dqlite we disabled several
checks around machine and unit agent version until migration could be
implemented.

This commit re-introduces these tests. As part of this work I have
introduced a new helper that will build up the set of expects for
modelagents based on the fake backend. This is a temporary measure until
we are off of Mongo.