feat: add mass precheck option to model agent
Adds two new functions to the model agent service for check if units and
machines in the models are all up and running to the target agent
version of the model.

This function was introduced to support model migrations. Currently we
loop over every machine and unit one by one hitting the database each
time. This is widly innuficient and we can perform this check as a bulk
call.