chore: remove exposed column from application table
This patch removes the exposed column from the application table because
it's not needed.
The method IsApplicationExposed() doesn't look at it (it checks if there
are any rows in the application_exposed_endpoint_space and
application_exposed_endpoint_cidr tables) and it is never set.

Temporarily, it was removed from the exported applications so it's not
migrated. This will be correctly implemented in a follow-up patch.