refactor: add current version to target agent
For machines and units we maintain target agent version tables for both
entities that signals what the current target agent version is for each
entity.

However this table was missing the entity name and also the current
reported agent version the entity is running. The reason for adding this
is so we can do one big bulk check on these entities during migration to
make sure that they are running the target agent version before
migration.

Currently this check is done in migration pre-checks asking the database
each time about singular entities. For a large complicated model this
check will be expensive when the question can be formulated in one query
inside of a single transaction.