fix: do not attempt to validate model defaults
Model defaults, when set are checked for validity by applying them to
the model config for the state's model.

This is not correct. We should not be using a config validator for a
model's cloud that may be different to another model to which we want to
apply the defaults.

Furthermore, it introduces a kind of deadlock whereby:
- Model default for legacy proxy settings are used at bootstrap.
- Every model created then has those settings.
- Setting the default for legacy proxy settings back to "" works.
- But now setting the new juju-* proxy values fails because we validate
  against the current model config.

Model defaults are used to snapshot new model config, not to fall back
to so we can rely on the model-level config validation, also applied at
creation to ensure we never have invalid config.