refactor(domain/relation/state): streamline test helpers usage
Helpers in relation_test add object from an UUID given as argument. This PR inverse the logic and
make all the helpers prefixed by "add" and inserting an actual entity in the db (something that have
a uuid) return the inserted UUID.

It reduce the clutter in the arrange phase of test and provide more flexibility.

- `domain/relation/state/relation_test.go`:
  - Replaced manual UUID generation and variable assignments like `GenRelationUUID` with test
   helper functions (`addCharm`, `addRelation`, and others) for clarity and consistency.
   - Removed unused imports and redundant declarations, simplifying the overall test structure.