chore: add space lookup for pre-4.0 exposed endpoints migration
This patch fixes a wrong behavior for pre-4.0 migrations, in which the
space IDs of the exposed endpoints (if any) when migrated to 4.0
wouldn't find the correct matching space because of the new UUID while
inserting them.
The fix is to do a lookup by name of each exposed endpoint to a space
and getting the space name from the description model spaces.