refactor: remove openstack provider upgrade checks
The openstack provider had implemented upgrade steps for controller
model upgrade in Juju 2.8.0. The check was to see if neutron networking
was supported.

We are well past 2.8.0 now and neutron is required for all new
controllers. We also don't support upgrade the controller anymore. This
check is not needed and doesn't need to be apart of model upgrading
anymore.