refactor: remove environ and upgrade from  model upgrader
This commit removes the environ and also the provider based model
upgrade steps from the model upgrader facade. These checks are not being
used anymore and the only provider that ever supported them was
openstack which was only for 2.8.0 upgrades.

This adds a lot of non needed complexity to model upgrades that doesn't
need to exist.