refactor: change api root model upgrade
This change affects how the model upgrade command works. Currently the
command logs into the controller with a controller root and passes as an
argument to the model upgrader facade the model uuid that needs to be
upgraded.

This breaks Juju patterns badly. Now with this change model upgrade
command logs into the facade as the model being upgraded.