refactor: removes caas broker from model manager
In previous commits we removed the ability for the facade to directly
make provider resources when new models are created. We did leave behind
the new caas broker func from when we used to create k8s resources in
the model manager facade.

This commit just removes this dep as we don't use it.