fix: enter scope test subordinate test
One of the helper functions in the test suite was not setting the
subordinate value in the charm metadata to true as expected. The table
had been set to immutable and becase of that it was setting it to true
when it should have been false.

This hid an actual error in the code, a conditional was checking if a
value was false when it should have been checking if it was true.

This change fixes that, and replaces the offending function with one
that just inserts, and does not update.