feat: implement device constraint state layer
This patch adds two state layer methods to insert and retrieve device constraints for a given application.
Also, as a bonus, the bson tags in the core/devices Constraint{} struct
were removed since they shouldn't have been in a core struct in the
first place (and mongo is being removed anyways). Also, the int64 for
the count field was replaced by an int for convenience (since it's
mapped to sql types at the state layer anyways).