refactor: upgrade validation steps
This commit starts to introduce a set of changes to the upgrade
validation steps we have for models and controllers. Specifically the
goal here is to start looking at ways to remove Mongo state out of these
validators.

In this commit we have removed model from the validators. To do this we
have removed the ModelMigration validator. This was done with the view
that it was the only use case that we still had using model. This check
will be added back in later in a domain service as part of upgrade.

To add the check back in the rest of model migration status will need to
be migrated to Dqlite.

The other use of model was for the validation blockers. When a
validation fails it will supply back a set of blockers to the caller on
why validation can't proceed. As part of the error message generated we
use the model owner and name to signify the model that has failed the
checks.

We don't need a whole model object to do this and instead we can rely on
the caller supply the model name to us for what they want to see in the
error message. This inverts the dependencies and makes the code much
easier to reason about.

The last change in this commit is to remove two validators that we had
around mongo replica set health and version. As Juju 4.0 is shifting off
of Mongo completely we will get rid of these checks now.

Removing the mongo validators allows us to remove the need for state
pool to the validators.