refactor: remove more uses of mongo model state
This commit attempts to remove more pieces from the model upgrader that
are relying on mongo state for accessing model information.

In this commit we were able to move the model type parameters for
finding a desired version over to using core types instead of state.

We are able to now use the model info service for model information
about the model being upgraded. As part of this we have also wired in
the current model's agent binary service.

However this still leaves a big hole when it comes to upgrading
controllers. To currently perform this task we still need state as this
check requires looping through every model in the controller and running
checks. To remove these last pieces we are going to have to develop new
domains for upgrading the controller and model that this facade can use.

This will allow us to remove the last of the mongo calls from the model
upgrader.